GDT almost works now

+++ kernel/includes/gdt.h
@@ -1,7 +1,9 @@
 #ifndef GDT_H
 #define GDT_H
 
-void add_gdt_entry(int i, unsigned int base_addr, unsigned int size, int flags);
+#include "stdint.h"
+
+void add_gdt_entry(int i, uint32_t base_addr, uint32_t size, uint16_t flags);
 void load_gdt();
 
 #endif /* GDT_H */
